Feat/#87 business

+++ src/main/java/org/ever/_4ever_be_gw/business/dto/analytics/ScaleDto.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class ScaleDto {
-    private Long min;
-    private Long max;
+    private BigDecimal min;
+    private BigDecimal max;
 }
